requirements.txt
langchain==0.1.9
streamlit==1.31.1
streamlit-chat==0.1.1
openai==1.12.0
faiss-cpu==1.7.4
PyPDF==4.0.2
tiktoken==0.6.0
python-dotenv==1.0.1
langchain-community==0.0.24
langchain-openai==0.0.8
langchain_mistralai==0.0.5
